chore(brand): remove dead persona proxy routes (#589)
Wave 2 persona→audience unification cleanup. Brand-service is deleting
its persona table + routes; audiences (human-service) fully replaced
personas and the lead→features→human runtime is live in prod. The
gateway's /v1/brands/:id/personas* proxies (list, create, suggest,
duplicate, avatar/regenerate, status) have zero callers across
dashboard/lead/campaign/features and would 404 once brand-service drops
them.

Remove all 6 brand persona proxy routes + their Zod/OpenAPI registrations
+ tests; regenerate openapi.json. Untouched: ICP suggest, brand-profile,
audiences (/v1/orgs/audiences/*), and features /persona-stats (a separate
features-service surface, not brand personas).
